perubahan ketiga

+++ application/controllers/lelang/Nota_penerimaan_lelang.php
@@ -8,6 +8,9 @@ class Nota_penerimaan_lelang extends CI_Controller
         is_logged_in();
         $this->load->model('Data_nota_model', 'nota');
         $this->load->model('Data_detail_model', 'detail');
+        $this->load->model('Ref_nota_model', 'refnota');
+        $this->load->model('Ref_satker_model', 'satker');
+        $this->load->model('Data_transaksi_model', 'transaksi');
     }
 
     public function index()
@@ -16,7 +19,7 @@ class Nota_penerimaan_lelang extends CI_Controller
 
         // setting halaman
         $config['base_url'] = base_url('lelang/nota-penerimaan-lelang/index/');
-        $config['total_rows'] = $this->nota->countApp(sesi()['kdsatker'], sesi()['tahun']);
+        $config['total_rows'] = $this->nota->countDetail(sesi()['kdsatker'], sesi()['tahun'], 'D', 'L', 0);
         $config['per_page'] = 10;
         $config["num_links"] = 3;
         $this->pagination->initialize($config);
@@ -30,9 +33,9 @@ class Nota_penerimaan_lelang extends CI_Controller
 
         // pilih tampilan data, semua atau berdasarkan pencarian
         if ($data['urut']) {
-            $data['nota'] = $this->nota->findApp(sesi()['kdsatker'], sesi()['tahun'], $data['urut'], $limit, $offset);
+            $data['nota'] = $this->nota->findDetail(sesi()['kdsatker'], sesi()['tahun'], 'D', 'L', 0, $data['urut'], $limit, $offset);
         } else {
-            $data['nota'] = $this->nota->getApp(sesi()['kdsatker'], sesi()['tahun'], $limit, $offset);
+            $data['nota'] = $this->nota->getDetail(sesi()['kdsatker'], sesi()['tahun'], 'D', 'L', 0, $limit, $offset);
         }
 
         $this->load->view('template/header', $data);
@@ -41,31 +44,38 @@ class Nota_penerimaan_lelang extends CI_Controller
         $this->load->view('template/footer');
     }
 
-    private $rules_nota = [
+    private $rules = [
         [
-            'field' => 'nominal',
-            'label' => 'Nominal',
-            'rules' => 'required|numeric'
+            'field' => 'kode',
+            'label' => 'kode',
+            'rules' => 'required'
         ]
     ];
 
     public function create()
     {
         $data['header'] = 'Nota Penerimaan Lelang';
+        $data['refnota'] = $this->refnota->getKegJns('L', 'D');
 
-        $validation = $this->form_validation->set_rules($this->rules_nota);
+        $validation = $this->form_validation->set_rules($this->rules);
         if ($validation->run()) {
+            $kode = htmlspecialchars($this->input->post('kode', true));
+            $kdn = substr($kode, 0, 2);
+            $ket = substr($kode, 2, strlen(trim($kode)) - 2);
             $data = [
                 'kdsatker' => sesi()['kdsatker'],
                 'bulan' => date('m'),
                 'tahun' => sesi()['tahun'],
-                'kdn' => '01',
-                'urut' =>  '0004',
-                'nominal' => htmlspecialchars($this->input->post('nominal', true)),
+                'kdn' => $kdn,
+                'urut' =>  urut_nota()['urut_nota'],
+                'jumlah' => 0,
+                'nominal' => 0,
                 'tanggal' => time(),
-                'rekening_id' => 1
+                'ket' => $ket,
+                'keg' => 'L'
             ];
             $this->nota->create($data);
+            $this->satker->updateKode(['urut_nota' => urut_nota()['urut_nota_next']], sesi()['kdsatker']);
             $this->session->set_flashdata('success', 'Data berhasil ditambah.');
             redirect('lelang/nota-penerimaan-lelang');
         }
@@ -76,36 +86,60 @@ class Nota_penerimaan_lelang extends CI_Controller
         $this->load->view('template/footer');
     }
 
-    public function update($id = null)
+    public function update($id = null, $jumlah = 1)
     {
-        $data['header'] = 'Nota Penerimaan Lelang';
-
         if (!isset($id)) show_404();
-        $data['nota'] = $this->nota->getApp(sesi()['kdsatker'], sesi()['tahun'], 0, null, $id);
 
-        $validation = $this->form_validation->set_rules($this->rules_nota);
-
-        if ($validation->run()) {
-            $data = [
-                'nominal' =>  htmlspecialchars($this->input->post('nominal', true))
-            ];
-            $this->nota->update($data, $id);
-            $this->session->set_flashdata('success', 'Data berhasil diubah.');
+        if ($jumlah > 0) {
+            $this->session->set_flashdata('danger', 'Data tidak dapat diubah, pada detail nota terdapat ' . $jumlah . ' transaksi yang belum dihapus.');
             redirect('lelang/nota-penerimaan-lelang');
-        }
+        } else {
+            $data['header'] = 'Nota Penerimaan Lelang';
+            $data['nota'] = $this->nota->getApp(sesi()['kdsatker'], sesi()['tahun'], 0, null, $id);
 
-        $this->load->view('template/header', $data);
-        $this->load->view('template/sidebar');
-        $this->load->view('lelang/nota_penerimaan_lelang/update', $data);
-        $this->load->view('template/footer');
+            $validation = $this->form_validation->set_rules($this->rules);
+
+            if ($validation->run()) {
+                $data = [
+                    'nominal' =>  htmlspecialchars($this->input->post('nominal', true))
+                ];
+                $this->nota->update($data, $id);
+                $this->session->set_flashdata('success', 'Data berhasil diubah.');
+                redirect('lelang/nota-penerimaan-lelang');
+            }
+
+            $this->load->view('template/header', $data);
+            $this->load->view('template/sidebar');
+            $this->load->view('lelang/nota_penerimaan_lelang/update', $data);
+            $this->load->view('template/footer');
+        }
     }
 
-    public function delete($id = null)
+    public function delete($id = null, $jumlah = 1)
     {
         if (!isset($id)) show_404();
 
-        if ($this->nota->delete($id)) {
-            $this->session->set_flashdata('danger', 'Data berhasil dihapus.');
+        if ($jumlah > 0) {
+            $this->session->set_flashdata('danger', 'Data tidak dapat dihapus, pada detail nota terdapat ' . $jumlah . ' transaksi yang belum dihapus.');
+            redirect('lelang/nota-penerimaan-lelang');
+        } else {
+            if ($this->nota->delete($id)) {
+                $this->session->set_flashdata('success', 'Data berhasil dihapus.');
+            }
+        }
+        redirect('lelang/nota-penerimaan-lelang');
+    }
+
+    public function process($id = null)
+    {
+        if (!isset($id)) show_404();
+
+        if ($this->nota->update(['sts' => 1, 'tgl_kirim' => time()], $id)) {
+            $detail = $this->detail->getNotaId($id);
+            foreach ($detail as $r) {
+                $this->transaksi->update(['nota_d_id' => $id], $r['transaksi_id']);
+            }
+            $this->session->set_flashdata('success', 'Data berhasil diproses.');
         }
         redirect('lelang/nota-penerimaan-lelang');
     }
@@ -148,45 +182,84 @@ class Nota_penerimaan_lelang extends CI_Controller
     {
         if (!isset($id)) show_404();
 
+        $detail = $this->detail->get($id);
+        $transaksi_id = $detail['transaksi_id'];
+
         if ($this->detail->delete($id)) {
-            $this->session->set_flashdata('danger', 'Data berhasil dihapus.');
+            $this->transaksi->update(['sts' => 0], $transaksi_id);
+            $jumlah = $this->detail->countNotaId($nota_id);
+            $nominal = $this->detail->sumNotaId($nota_id)['nominal'];
+            $this->nota->update(['jumlah' => $jumlah, 'nominal' => $nominal], $nota_id);
+            $this->session->set_flashdata('success', 'Data berhasil dihapus.');
         }
         redirect('lelang/nota-penerimaan-lelang/detail/' . $nota_id . '');
     }
 
-    private $rules_detail = [
-        [
-            'field' => 'nominal',
-            'label' => 'Nominal',
-            'rules' => 'required|numeric'
-        ]
-    ];
-
     public function create_detail($nota_id = null)
     {
         if (!isset($nota_id)) show_404();
-        $data['header'] = 'Nota Penerimaan Lelang';
 
-        $validation = $this->form_validation->set_rules($this->rules);
-        if ($validation->run()) {
-            $data = [
-                'kdsatker' => sesi()['kdsatker'],
-                'bulan' => date('m'),
-                'tahun' => sesi()['tahun'],
-                'kdn' => '01',
-                'urut' =>  '0004',
-                'nominal' => htmlspecialchars($this->input->post('nominal', true)),
-                'tanggal' => time(),
-                'rekening_id' => 1
-            ];
-            $this->nota->create($data);
-            $this->session->set_flashdata('success', 'Data berhasil ditambah.');
-            redirect('lelang/nota-penerimaan-lelang');
+        $data['header'] = 'Nota Penerimaan Lelang';
+        $data['nota_id'] = $nota_id;
+
+        $kode = $this->nota->get($nota_id)['kdn'];
+        $refnota = $this->refnota->getKode($kode);
+        $kdk = $refnota['kdk'];
+        $kdj = $refnota['kdj'];
+
+        // setting halaman
+        $config['base_url'] = base_url('lelang/nota-penerimaan-lelang/create-detail/' . $nota_id . '/');
+        $config['total_rows'] = $this->transaksi->countNota(sesi()['kdsatker'], sesi()['tahun'], $kdk, $kdj, 'D', 'L', 0);
+        $config['per_page'] = 10;
+        $config["num_links"] = 3;
+        $this->pagination->initialize($config);
+        $data['pagination'] = $this->pagination->create_links();
+        $data['page'] = $this->uri->segment(5) ? $this->uri->segment(5) : 0;
+        $limit = $config["per_page"];
+        $offset = $data['page'];
+
+        // menangkap pencarian jika ada
+        $data['uraian'] = $this->input->post('uraian');
+
+        // pilih tampilan data, semua atau berdasarkan pencarian
+        if ($data['uraian']) {
+            $data['transaksi'] = $this->transaksi->findNota(sesi()['kdsatker'], sesi()['tahun'], $kdk, $kdj, 'D', 'L', 0, $data['uraian'], $limit, $offset);
+        } else {
+            $data['transaksi'] = $this->transaksi->getNota(sesi()['kdsatker'], sesi()['tahun'], $kdk, $kdj, 'D', 'L', 0, $limit, $offset);
         }
 
         $this->load->view('template/header', $data);
         $this->load->view('template/sidebar');
-        $this->load->view('lelang/nota_penerimaan_lelang/create');
+        $this->load->view('lelang/nota_penerimaan_lelang/create_detail', $data);
         $this->load->view('template/footer');
+    }
+
+    public function add_detail($nota_id = null, $id = null)
+    {
+        if (!isset($nota_id)) show_404();
+        if (!isset($id)) show_404();
+
+        if ($this->transaksi->update(['sts' => 1], $id)) {
+            $t = $this->transaksi->get($id);
+            $data = [
+                'kdsatker' => $t['kdsatker'],
+                'bulan' => $t['bulan'],
+                'tahun' => $t['tahun'],
+                'kdk' => $t['kdk'],
+                'kdj' =>  $t['kdj'],
+                'urut' => $t['urut'],
+                'nominal' => $t['nominal'],
+                'nota_id' => $nota_id,
+                'transaksi_id' => $t['id'],
+                'sts' => 0
+            ];
+            $this->detail->create($data);
+
+            $jumlah = $this->detail->countNotaId($nota_id);
+            $nominal = $this->detail->sumNotaId($nota_id)['nominal'];
+            $this->nota->update(['jumlah' => $jumlah, 'nominal' => $nominal], $nota_id);
+            $this->session->set_flashdata('success', 'Data berhasil dipilih.');
+        }
+        redirect('lelang/nota-penerimaan-lelang/create-detail/' . $nota_id . '');
     }
 }
